requirements.txt
matplotlib
imagehash
pillow
